applications/foldrun/src/foldrun-analysis-job/requirements.txt
google-cloud-storage>=3.0.0
google-cloud-aiplatform>=1.85.0
google-genai>=1.0.0
numpy>=2.2.3
matplotlib>=3.10.0
seaborn>=0.13.2
pyyaml>=6.0
jax[cpu]==0.4.38
dm-tree>=0.1.9
